Tue, Oct 24, 2023, 3:32 PM +05:30

+++ hours/hours.c
@@ -38,7 +38,7 @@ float calc_hours(int hours[], int weeks, char output)
 
     if (output == 'A')
     {
-     int average = total / weeks ;
+     float average = (float) total / weeks ;
      return average;
      }
